feat: add Google's SafeBrowsing service
Google SafeBrowsing allows our service to scans URLs against a database
of URLs reported to be prejudicial to users. This is required to ensure
that our service is not being used with evil intentions, and
consequentially being reported as malware in other URL databases.

When checking for a URL, our service will simply return a `true` or
`false` response, depending if the any of the submitted URLs are
reported as prejudicial or not.

If the SafeBrowsing request fails, the function will return `true` to
avoid blocking the user from submitting a URL.