feat(database): add sequelize
Use [sequelize v5][] as our ORM.
It provides out-of-box support for most common relational databases (pg,
mysql, sqlite).

We take advantage of this support by running unit tests against an
in-memory sqlite database and our integration tests against a full
Postgres database.

The choice for sequelize, is because, at the moment is the most popular
ORM library in active development.

[sequelize v5]: https://sequelize.org/v5/