feat(config): add database configs
We are already using [node-config][] to manage our configurations.
However [sequelize][] requires its own .json or .js file with just their
options.

In order to support that we've created the
`config/readonly-database-config.js` with the sole purpose of importing
the database configurations from node-config and export them again in a
format that sequelize understands.

[node-config]: https://lorenwest.github.io/node-config/
[sequelize]: https://sequelize.org/master/manual/migrations.html#configuration