build(docker): create docker image for web service
Use docker to create an isolated enviroment for building our project,
and take advantage of its [multi-stage][] capabilities to create a final
image with only the production dependencies.

As a base layer we are using the [oficial node][] image with the alpine
flavor in order to keep our images as small as possible.

[multi-stage]: https://docs.docker.com/develop/develop-images/multistage-build/
[offical node]: https://hub.docker.com/_/node/