feat(service): add password service
[Argon2][] is a key derivation function that was selected as the
winner of the Password Hashing Competition in July 2015.

It has been 5 years since the launch of Argon2. It's been proven stable
and secure so far.

We are using the [node-argon2][] implementation.

NOTE: We still need to figure out the best performance settings to
provide an acceptable security without consuming too much resources.

[Argon2]: https://en.wikipedia.org/wiki/Argon2
[node-argon2]: https://github.com/ranisalt/node-argon2